feat: basic web implementation for native stack
